% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayo_css_files.R
\name{mayo_css_files}
\alias{mayo_css_files}
\title{List available CSS files}
\usage{
mayo_css_files()
}
\description{
Print a list of availabe CSS files included in mayotheme
}
\examples{
mayo_css_files()
}
