% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayocolors.R
\docType{data}
\name{mayocolors}
\alias{mayocolors}
\title{Mayo Clinic Colors}
\format{A list}
\usage{
mayocolors
}
\description{
Primary, accent, and special use colors for Mayo Clinic
}
\examples{
mayocolors
}
\keyword{datasets}
