% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mayo_logo.R
\name{use_mayo_logo}
\alias{use_mayo_logo}
\title{Use Mayo Logo}
\usage{
use_mayo_logo(logo = c("black", "secondary"), data_uri = FALSE)
}
\arguments{
\item{logo}{Logo type}

\item{data_uri}{Logical.  Should a data uri be returned?
If FALSE, file location is returned.}
}
\description{
Returns location or data uri of a Mayo logo
}
